step count bugs fixes

+++ lib/tracker_screens/diet_tracker_screen/sections/water_goal.dart
@@ -51,7 +51,7 @@ class _WaterGoal extends State<WaterGoal> {
   }
 
   double getPointerValue() {
-    return ((int.parse(currentTakenWater) / int.parse(waterGoal)) * 100);
+    return ((double.parse(currentTakenWater) / double.parse(waterGoal)) * 100);
   }
 
   @override
